pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "user-feedback-clustering"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "numpy>=1.25.0",
    "typer>=0.9.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
feedback-cluster = "feedback_clustering.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
